chat-app/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-socketio==5.11.4
motor==3.6.0
pydantic==2.10.2
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.1
email-validator>=2.0.0
boto3==1.35.76
python-multipart==0.0.20
apscheduler==3.10.4
httpx==0.27.0